Cleanly convert register readings to an integer and store it as such in the DB as well. 2 tests fail, I don't know why. I'd tend to remove the call to Reading::Single#corrected_value, but not sure.
     # ./spec/requests/admin/reading_spec.rb:175:in `block (4 levels) in <top (required)>'
     # ./spec/requests/admin/reading_spec.rb:144:in `block (4 levels) in <top (required)>'

     Failure/Error: expect(json.to_yaml).to eq(reading_json.to_yaml)       (compared using ==)

       Diff:
       -value: 1234
       +value: 1234.0

+++ db/structure.sql
@@ -1696,8 +1696,8 @@ CREATE TABLE persons_roles (
 
 CREATE TABLE readings (
     id integer NOT NULL,
-    raw_value double precision NOT NULL,
-    value double precision NOT NULL,
+    raw_value integer NOT NULL,
+    value integer NOT NULL,
     comment character varying(256),
     date date,
     created_at timestamp without time zone NOT NULL,
